Revert "numademo: fix error on 32bit system"
Revert due to #135 -- cannot break API for programs

This reverts commit a7c4bc790a191d3e42b63850b409c1a72b75a4e1.

+++ libnuma.c
@@ -830,10 +830,10 @@ long long numa_node_size64(int node, long long *freep)
 
 make_internal_alias(numa_node_size64);
 
-long long numa_node_size(int node, long long *freep)
+long numa_node_size(int node, long *freep)
 {
 	long long f2;
-	long long sz = numa_node_size64_int(node, &f2);
+	long sz = numa_node_size64_int(node, &f2);
 	if (freep)
 		*freep = f2;
 	return sz;
